Finish server side REST implementation
